GroceryShoppingSystem/src/java/DTO/StoreDTO.java
package DTO;


public class StoreDTO 
{
      private int id;      
      private String addressLine1;
      private String addressLine2;      
      private String postCode;

    public StoreDTO(int id, String addressLine1, String addressLine2, String postCode) 
    {
        this.id = id;
        this.addressLine1 = addressLine1;
        this.addressLine2 = addressLine2;
        this.postCode = postCode;
    }

    public int getId() {
        return id;
    }

    public void setId(int id) {
        this.id = id;
    }

    public String getAddressLine1() {
        return addressLine1;
    }

    public void setAddressLine1(String addressLine1) {
        this.addressLine1 = addressLine1;
    }

    public String getAddressLine2() {
        return addressLine2;
    }

    public void setAddressLine2(String addressLine2) {
        this.addressLine2 = addressLine2;
    }

    public String getPostCode() {
        return postCode;
    }

    public void setPostCode(String postCode) {
        this.postCode = postCode;
    }      
      
}
